switch to using django’s raw queries for inserts and deletions
No need to carry another dependency, we can achieve the same with Django’s already available functions.

Instead of using a single cursor, we create a new one and don’t need to have an additional variable when the performance cost is negligible.

Instead of using Django’s `.delete()` which made an additional select query, we run a raw `DELETE WHERE` query.